requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

opencv-python==4.8.1.78
mediapipe==0.10.7
numpy==1.24.3
Pillow==10.0.1

pydantic==2.4.2
python-dotenv==1.0.0
